Release 3 (#60)
* [IN-117] Local installation of the project (documentation) (#41)

* docs: enhancing docs for local installation

* chore: relational database diagram

* chore: remove crypto

* chore: fake spaces

* [IN-120] Final report (#45)

* [IN-79] GANTT Chart + forecasted planning on one year (#43)

* [IN-116] UX innovation propositions (#47)

* [IN-78] Call for tenders response (#46)

* chore: response call for tenders

* fix: remove duplicate report

* chore: update final report

* [IN-118] API and deployment process documentation (#44)

* chore: api documentation

* chore: update README.md

* chore: update README.md + deployment process

* chore: remade deployment process

* fix: spelling errors

* fix: another spelling error

* style: spaces

* chore: update main README

* chore: update README

* [IN-65] Delete any of my own original content as a MINTER (#42)

* feat: fetch original contents and perform delete action.

* feat: translation

* refactor : translations and refactoring accoring to PR review.

* refactor: to use useRelativeTime to display the og content created time.

* [IN-107] Replace Number() use by parseInt (#49)

* build(deps): bump ws in the npm_and_yarn group across 1 directory (#50)

Bumps the npm_and_yarn group with 1 update in the / directory: [ws](https://github.com/websockets/ws).

Updates `ws` from 8.17.0 to 8.17.1
- [Release notes](https://github.com/websockets/ws/releases)
- [Commits](https://github.com/websockets/ws/compare/8.17.0...8.17.1)

---
updated-dependencies:
- dependency-name: ws
  dependency-type: direct:production
  dependency-group: npm_and_yarn
...


* build(deps): bump braces in the npm_and_yarn group across 1 directory (#51)

Bumps the npm_and_yarn group with 1 update in the / directory: [braces](https://github.com/micromatch/braces).

Updates `braces` from 3.0.2 to 3.0.3
- [Changelog](https://github.com/micromatch/braces/blob/master/CHANGELOG.md)
- [Commits](https://github.com/micromatch/braces/compare/3.0.2...3.0.3)

---
updated-dependencies:
- dependency-name: braces
  dependency-type: indirect
  dependency-group: npm_and_yarn
...


* [IN-60] Dashboard with analytics (backoffice) (#48)

* chore: chartjs dependency + api structure

* feat: working nfts (mint) count api

* fix: misplace use client + init prisma query files

* chore: count for mints + comments by nfts + mints plus comments by nfts + count active minters

* fix: cve found in braces package

* chore: removed unreachable metrics for now

* feat: working feature (need to add chart)

* refactor: metrics charts

* feat: working feature (comments + comments by nft + active minters)

* chore: query => calculateSince

* [IN-95] Middleware implementation from Auth.js (#52)

* chore: update middleware

* chore: dependencies

* [IN-114] Refactor LinkButton component (#56)

* [IN-35] Share a link to a specific NFT as a MINTER (#55)

* feat: web sharing functionality to specific nft using react-web-share library.

* feat: add translation

* feat: add share feature inside an NFT page and in search NFT card.

* [IN-63] Disable MINTER account (backoffice) (#53)

* feat: add isActivated column on User table

* feat: disable minter account

* feat: display error if account is disabled on login

* refactor: create hook to centralize minter actions

* feat: translate user not activated message

* feat(minter): protect delete endpoint

* chore: apply PR review requested changes

* [IN-105] Refactor and enhance source code (#57)

* chore: server and client separation

* chore: axios instance

* chore: axios timeout

* chore: handler before jsx

* chore: removed useless catch

* fix: return type

* chore: removed all chaining style from application

* refactor: translation schema split

* fix: translation errors and other bad practices

* style: format code from sign up logic

* refactor: change native html input to shadcn input

* style: original contents

* refactor: with error handling api wrapper

* style: format code

* feat: enhance UI (error message + success message)

* [IN-123] Hotfix: User ID => Minter ID (#58)

* fix: user id => minter id

* feat: user id => minter id for nft details

* [IN-124] Hotfix - English translation + update dislike after previous fix

* Review - Sprint 3

* feat: sprint report 3

* style: spaces

* fix: archi diagram

---------

+++ src/app/api/minters/[id]/route.ts
@@ -1,16 +1,38 @@
-import deleteMinter from "@/lib/query/minters/deleteMinter"
-import getMinterById from "@/lib/query/minters/getMinterById"
-import { StatusCodes } from "http-status-codes"
+import { withErrorHandling } from "@/lib/helpers/apiWrapper"
+import deleteMinter from "@/lib/query/server/minters/deleteMinter"
+import getMinterById from "@/lib/query/server/minters/getMinterById"
+import isAdmin from "@/lib/utils/auth/isAdmin"
+import isAuthenticated from "@/lib/utils/auth/isAuthenticated"
+import isCurrentUser from "@/lib/utils/auth/isCurrentUser"
+import { ReasonPhrases, StatusCodes } from "http-status-codes"
 
-export async function GET(req: Request, { params }: { params: { id: string } }) {
+export const GET = withErrorHandling(handleGet)
+
+export const DELETE = withErrorHandling(handleDelete)
+
+async function handleGet(req: Request, { params }: { params: { id: string } }) {
   const { id } = params
-  const minter = await getMinterById(Number(id))
+  const minter = await getMinterById(parseInt(id, 10))
 
   return Response.json(minter)
 }
 
-export async function DELETE(req: Request, { params }: { params: { id: string } }) {
+async function handleDelete(req: Request, { params }: { params: { id: string } }) {
+  const session = await isAuthenticated()
+
+  if (!session) {
+    return Response.json(
+      { message: ReasonPhrases.UNAUTHORIZED },
+      { status: StatusCodes.UNAUTHORIZED }
+    )
+  }
+
   const { id } = params
+
+  if (!isCurrentUser(id, session) && !isAdmin(session)) {
+    return Response.json({ message: ReasonPhrases.FORBIDDEN }, { status: StatusCodes.FORBIDDEN })
+  }
+
   await deleteMinter(id)
 
   return new Response(null, { status: StatusCodes.NO_CONTENT })
